fix error test
